test: split the test file and embed it using the include macro.
